add Pry::Helpers::Colors, and #paint function.
